chore: Move `const` from `js` to `common` folder
