--- src/js/const/error-codes.const.ts
@@ -1,9 +1,0 @@
-export const API_ERROR_CODES = Object.freeze({
-  // generic
-  tooManyRequests: 'TOO_MANY_REQUESTS',
-
-  // monobank
-  monobankUserNotPaired: 'MONOBANK_USER_NOT_PAIRED',
-  monobankUserAlreadyConnected: 'MONOBANK_USER_ALREADY_CONNECTED',
-  monobankUserNotExist: 'MONOBANK_USER_NOT_EXIST',
-});
